+++ src/Divider/Divider.tsx
@@ -1,12 +1,13 @@
-import { useSeparator, SeparatorProps } from 'react-aria';
+import { useSeparator } from 'react-aria';
+import { DividerProps } from './types';
 import { Root } from './styles';
 
-export function Divider(props: SeparatorProps) {
-  const { orientation } = props;
+export function Divider(props: DividerProps) {
   const { separatorProps } = useSeparator(props);
+  const combinedProps = { ...props, ...separatorProps };
 
   // eslint-disable-next-line react/jsx-props-no-spreading
-  return <Root {...separatorProps} orientation={orientation} />;
+  return <Root {...combinedProps} />;
 }
 
 export default Divider;