+++ src/Divider/styles.ts
@@ -1,14 +1,18 @@
 import { styled } from 'styled-components';
-import { SeparatorProps } from 'react-aria';
 import { defaultTheme } from '../defaultTheme';
 
-export const Root = styled.div<{ orientation: SeparatorProps['orientation'] }>`
-  width: ${({ orientation }) => (orientation === 'vertical' ? '1px' : '100%')};
-  height: ${({ orientation }) => (orientation === 'vertical' ? 'auto' : '1px')};
-  margin: ${({ orientation }) =>
-    orientation === 'vertical' ? '0 1rem' : '1rem 0'};
+export const Root = styled.div`
+  width: 100%;
+  height: 1px;
+  margin: 1rem 0;
   background: ${({ theme }) =>
     theme?.colors?.slate?.[200] ?? defaultTheme.colors.slate[200]};
+
+  &[aria-orientation='vertical'] {
+    width: 1px;
+    height: 100%;
+    margin: 0 1rem;
+  }
 `;
 
 export default Root;