+++ src/Divider/types.ts
@@ -1,0 +1,5 @@
+import { SeparatorProps } from 'react-aria';
+
+export type DividerProps = SeparatorProps & {
+  className?: string;
+};